Add a 3-stage PP config
Pipelining is unique in that there is no need to stick to power-of-2
numbers of stages, and there maybe reasons an odd number is optimal
depending on how you divide up your cluster.

Anyway, I use this for validation of the 1f1b schedule in a slightly-more-complicated
than 2-stage but simpler than 4-stage setup.

seems to run fine, if run with an even batch size
(`--training.batch_size 12`)

ghstack-source-id: 72d366936cb4fd10de6c3a3ccea7f6d7e5089dfd
Pull Request resolved: https://github.com/pytorch/torchtitan/pull/345